requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
autopep8==2.3.1
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
distro==1.9.0
Django==4.2
django-extensions==3.2.3
django-redis==5.4.0
django-seed==0.3.1
django-silk==5.2.0
django-taggit==6.0.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
docstring-parser==0.15
drf-spectacular==0.27.2
exceptiongroup==1.2.2
Faker==28.0.0
Flask==3.0.3
Flask-Cors==4.0.1
Flask-SQLAlchemy==3.1.1
frozenlist==1.4.1
gprof2dot==2024.6.6
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
importlib_metadata==8.0.0
inflection==0.5.1
instructor==0.6.8
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.5.0
jsonify==0.5
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
multidict==6.0.5
openai==1.42.0
pillow==10.4.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycodestyle==2.12.1
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
PyJWT==2.9.0
pyreadline3==3.4.1
python-dateutil==2.9.0.post0
python-pptx==1.0.2
PyYAML==6.0.2
redis==5.0.8
referencing==0.35.1
requests==2.32.3
rich==13.8.0
rpds-py==0.20.0
shell_gpt==1.4.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.31
sqlparse==0.5.1
tenacity==8.5.0
tomli==2.0.1
toposort==1.10
tqdm==4.66.5
typer==0.9.4
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
Werkzeug==3.0.3
XlsxWriter==3.2.0
yarl==1.9.4
zipp==3.19.2
